nav [nfc]: Move HomePage to a new home
