+++ lib/widgets/login.dart
@@ -14,8 +14,8 @@ import '../generated/l10n/zulip_localizations.dart';
 import '../log.dart';
 import '../model/binding.dart';
 import '../model/store.dart';
-import 'app.dart';
 import 'dialog.dart';
+import 'home.dart';
 import 'input.dart';
 import 'page.dart';
 import 'store.dart';
